pyproject.toml
[project]
name = "python-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
